pyproject.toml
[project]
name = "ledger-load-test"
version = "0.1.0"
description = "Load testing suite for Nabu Ledger gRPC API (Flutterwave RFQ PoC)"
requires-python = ">=3.10"
dependencies = [
    "locust>=2.20",
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.25",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "generated"]
